requirements.txt
langchain~=0.0.202
openai~=0.27.8
unstructured~=0.7.6
faiss-cpu~=1.7.4
gradio~=3.35.2
pypdf~=3.10.0
tiktoken~=0.4.0
python-dotenv~=1.0.0
